Update README.md
The concept of the metaverse has been speedily changing from a futuristic idea to an immersive reality. As a leading provider in the digital frontier, Blockcoaster is at the lead of metaverse game development, manipulating up-to-date  technology to create expansive virtual worlds. Our expertise in blockchain, virtual reality (VR), augmented reality (AR), and game design ensures the creation of interactive and engaging experiences. In this blog, we explore the essentials of developing games in the metaverse and how Blockcoaster gives to this evolving landscape.